# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "BoundingBox": {
        "properties": {
          "height": {
            "title": "Height",
            "type": "number"
          },
          "width": {
            "title": "Width",
            "type": "number"
          },
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "BoundingBox",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "diagnostic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnostic Id"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "retryable",
          "diagnostic_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "The one error shape, for the generated contract (the handlers below produce it).",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "Forget": {
        "properties": {
          "expected_revision": {
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "Forget",
        "type": "object"
      },
      "ForgetPending": {
        "description": "Forgetting happened; part of the physical cleanup could not finish and is retried.",
        "properties": {
          "outstanding": {
            "items": {
              "type": "string"
            },
            "title": "Outstanding",
            "type": "array"
          }
        },
        "required": [
          "outstanding"
        ],
        "title": "ForgetPending",
        "type": "object"
      },
...
      "IdentitySummary": {
        "properties": {
          "activated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "occurrence_count": {
            "title": "Occurrence Count",
            "type": "integer"
          },
          "person": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "representative_observation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ObservationBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "source_count": {
            "title": "Source Count",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "revision",
          "person",
          "representative_observation",
          "occurrence_count",
          "source_count",
          "created_at",
          "activated_at"
        ],
        "title": "IdentitySummary",
        "type": "object"
      },
...
      "MediaReference": {
        "description": "Media is referenced, never embedded or located by a path (API section 22).",
        "properties": {
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "url",
          "content_type"
        ],
        "title": "MediaReference",
        "type": "object"
      },
      "Merge": {
        "properties": {
          "identities": {
            "items": {
              "$ref": "#/components/schemas/MergeMember"
            },
            "title": "Identities",
            "type": "array"
          },
          "preferred_identity_id": {
            "format": "uuid",
            "title": "Preferred Identity Id",
            "type": "string"
          }
        },
        "required": [
          "identities",
          "preferred_identity_id"
        ],
        "title": "Merge",
        "type": "object"
      },
      "MergeMember": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "revision"
        ],
        "title": "MergeMember",
        "type": "object"
      },
...
      "ObservationBrief": {
        "properties": {
          "bounding_box": {
            "$ref": "#/components/schemas/BoundingBox"
          },
          "face_crop": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_id",
          "bounding_box",
          "face_crop"
        ],
        "title": "ObservationBrief",
        "type": "object"
      },
...
      "PersonReference": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "display_name",
          "revision"
        ],
        "title": "PersonReference",
        "type": "object"
      },
...
      "Split": {
        "properties": {
          "expected_revision": {
            "title": "Expected Revision",
            "type": "integer"
          },
          "occurrence_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Occurrence Ids",
            "type": "array"
          }
        },
        "required": [
          "occurrence_ids",
          "expected_revision"
        ],
        "title": "Split",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FaceIdentify local API",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/identities/merge": {
      "post": {
        "operationId": "merge",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Merge"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentitySummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The one error shape."
          }
        },
        "summary": "Merge",
        "tags": [
          "identities"
        ]
      }
    },
...
    "/api/v1/identities/{identity_id}/forget": {
      "post": {
        "description": "Forget an identity (API section 8.3): the only operation that removes biometric memory. Its\nvectors are erased and its faces stop resolving to anyone; the media stays and so does any\nPerson it was linked to, with their name. A repeat is harmless.",
        "operationId": "forget",
        "parameters": [
          {
            "in": "path",
            "name": "identity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Forget"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgetPending"
                }
              }
            },
            "description": "Accepted"
          },
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The one error shape."
          }
        },
        "summary": "Forget",
        "tags": [
          "identities"
        ]
      }
    },
...
    "/api/v1/identities/{identity_id}/split": {
      "post": {
        "operationId": "split",
        "parameters": [
          {
            "in": "path",
            "name": "identity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Split"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentitySummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The one error shape."
          }
        },
        "summary": "Split",
        "tags": [
          "identities"
        ]
      }
    },
...
  }
}